project/requirements.txt
pytils==0.3
psycopg2==2.7.3.2
pillow==4.3.0
ipython==6.2.1
django==1.11.6
django-imagekit==4.0.2